backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.8.0
huggingface-hub==0.20.3
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
